requirements.txt
fastapi
uvicorn
pydantic
jinja2
markdown
python-multipart
python-dotenv
langchain-ollama